+++ scripts/jestSetup.js
@@ -17,3 +17,13 @@ global.IntersectionObserver = class {
 };
 
 global.__TARGET__ = 'web';
+
+const React = require('react');
+const ReactDOM = require('react-dom');
+
+const originalRender = ReactDOM.render;
+
+ReactDOM.render = (element, container, callback) => {
+  element = React.createElement(React.StrictMode, [], element);
+  return originalRender(element, container, callback);
+};